+++ app/src/main/java/de/welthungerhilfe/cgm/scanner/datasource/dao/PersonDao.java
@@ -62,8 +62,8 @@ public interface PersonDao {
     LiveData<List<Person>> getResultPerson(SupportSQLiteQuery query);
 
     @Query("SELECT COUNT(id) FROM " + TABLE_PERSON + " WHERE createdBy=:email")
-    int getOwnPersonCount(String email);
+    long getOwnPersonCount(String email);
 
     @Query("SELECT COUNT(id) FROM " + TABLE_PERSON)
-    int getTotalPersonCount();
+    long getTotalPersonCount();
 }
